Add authUpdateAttempted field to ServerMessage::AuthError (#34500)
We can hit `AuthError`s in two ways -- the existing auth expires or otherwise becomes invalid, or the client sends updated auth in an `Authenticate` message, which is not valid.

When the latter happens, we want to signal to the client that what it sent was not valid, and it should consider itself signed out and not try sending the same auth again.

However, this currently looks identical to the former, which can cause race conditions where the client interprets an `AuthError` meaning "the existing auth has expired" as "the auth you just sent me is invalid" (when in fact the auth sent over is totally valid).

To convey this to the client, I added a boolean `authUpdateAttempted` (naming suggestion welcome) + a new `ErrorMetadata` to capture this precise case.

I think the most correct way of doing this would be to condition sending `authUpdateAttempted` on the client version so any clients that are strict around new fields won't fail when this is added (looks like the Rust client would fail on extra fields since I believe that's the default for the derived `Deserialize` on `ServerMessageJson`)

GitOrigin-RevId: d553daad45688099a5aab5e2b142f77274a07698