Fix for a panic that could occur in a subscribe call
If a query was subscribed to multiple times by the same client, an unsubscribe followed by a subsequent subscribe (if in the right/wrong order) would trigger bad internal state and a panic.

This was due to a using a value that was incremented and decremented as part of the SubscriberId. That could lead to a SubscriberId for a new subscription matching a previous one, which is invalid state.