convex-rs PR 10: Fix for a panic that could occur in a subscribe call (#41227)
If a query was subscribed to multiple times by the same client, an unsubscribe followed by a subsequent subscribe (if in the right/wrong order) would trigger bad internal state and a panic.

This was due to a using a value that was incremented and decremented as part of the SubscriberId. That could lead to a SubscriberId for a new subscription matching a previous one, which is invalid state.

Co-authored-by: Christian Wyglendowski <christian@dowski.com>
GitOrigin-RevId: 04150d72ffba18410d332086d32d18ce33a7c67b